Emplement StartLeague Service

+++ YuGiTournament.Api/Services/LeagueResetService.cs
@@ -1,6 +1,11 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using YuGiTournament.Api.Services.Abstractions;
 using YuGiTournament.Api.Abstractions;
+using YuGiTournament.Api.ApiResponses;
+using YuGiTournament.Api.Models;
+using YuGiTournament.Api.DTOs;
+using System.Numerics;
+using Azure;
 
 namespace YuGiTournament.Api.Services
 {
@@ -13,17 +18,81 @@ namespace YuGiTournament.Api.Services
             _unitOfWork = unitOfWork;
         }
 
-        public async Task ResetLeagueAsync()
+        public async Task<LeagueResponse> GetCurrentLeague()
         {
-            //await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DELETE FROM Matches");
-            //await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DELETE FROM MatchRounds");
-            //await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DELETE FROM Players");
-            await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DELETE FROM Messages");
+            var league = await _unitOfWork.GetRepository<LeagueId>().Find(x => x.IsFinished == false).FirstOrDefaultAsync();
 
-            //await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DBCC CHECKIDENT ('Matches', RESEED, 0)");
-            //await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DBCC CHECKIDENT ('MatchRounds', RESEED, 0)");
-            //await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DBCC CHECKIDENT ('Players', RESEED, 0)");
-            await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DBCC CHECKIDENT ('Messages', RESEED, 0)");
+            if (league == null)
+                return new LeagueResponse { Response = new ApiResponse(false, "لا يوجد دوري حاليا"), League = null };
+
+            var response = new ApiResponse(true, "لا يوجد دوري حاليا");
+            return new LeagueResponse { Response = response, League = league };
+        }
+
+        public async Task<ApiResponse> ResetLeagueAsync(int leagueId)
+        {
+            var dbContext = _unitOfWork.GetDbContext();
+            using var transaction = await dbContext.Database.BeginTransactionAsync();
+
+            try
+            {
+
+                var league = _unitOfWork.GetRepository<LeagueId>().Find(x => x.Id == leagueId).FirstOrDefault();
+                if (league == null)
+                    return new ApiResponse(false, "الدوري ده مش موجود");
+
+                if (league.IsFinished == true)
+                    return new ApiResponse(false, "هذا الدوري منتهي بالفعل");
+                else
+                    league.IsFinished = false;
+
+
+                await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DELETE FROM Messages");
+                await _unitOfWork.GetDbContext().Database.ExecuteSqlRawAsync("DBCC CHECKIDENT ('Messages', RESEED, 0)");
+
+
+                await dbContext.Database.ExecuteSqlRawAsync(
+                    "UPDATE Matches SET IsDeleted = 1 WHERE LeagueNumber = {0}", leagueId);
+                await dbContext.Database.ExecuteSqlRawAsync(
+                    "UPDATE MatchRounds SET IsDeleted = 1 WHERE LeagueNumber = {0}", leagueId);
+                await dbContext.Database.ExecuteSqlRawAsync(
+                    "UPDATE Players SET IsDeleted = 1 WHERE LeagueNumber = {0}", leagueId);
+
+
+                await transaction.CommitAsync();
+            }
+            catch (Exception)
+            {
+                await transaction.RollbackAsync();
+                return new ApiResponse(false, "حدث خطأ اثناء انهاء الدوري");
+            }
+
+            return new ApiResponse(true, "تم انهاء الدوري");
+        }
+
+        public async Task<ApiResponse> StartLeagueAsync(StartLeagueDto newLeague)
+        {
+            var existLeague = await _unitOfWork.GetRepository<LeagueId>()
+                .Find(x => x.IsFinished == false)
+                .FirstOrDefaultAsync();
+            if (existLeague != null)
+                return new ApiResponse(false, $"هناك دوري حالي بالفعل");
+
+            var league = new LeagueId
+            {
+                Description = newLeague.Description,
+                Name = newLeague.Name,
+                TypeOfLeague = newLeague.TypeOfLeague,
+                IsFinished = false
+            };
+
+            var result = await _unitOfWork.GetRepository<LeagueId>().AddAsync(league);
+            if (!result)
+                return new ApiResponse(false, $"حدث خطأ اثناء حفظ الدوري");
+
+            await _unitOfWork.SaveChangesAsync();
+
+            return new ApiResponse(true, $"تم انشاء الدوري بنجاح");
         }
     }
 }
